fix lai hoa don cho

+++ src/Service/Impl/HoaDonService.java
@@ -68,8 +68,8 @@ public class HoaDonService implements IHoaDonService{
 //        hdctrepos.add(hd2, maSps);
     }
     
-    public void updateTrangThaiHoaDon(String maHDCT, Integer TrangThaiHD, Double TongTien, String mahd){
-        hoadonrepos.updateTrangThaiHoaDon(maHDCT, TrangThaiHD, TongTien, mahd);
+    public void updateTrangThaiHoaDon(String maHDCT,Integer TrangThaiHD, Float TongTien, String mahd){
+        hoadonrepos.updateTrangThaiHoaDon(maHDCT,TrangThaiHD, TongTien, mahd);
     }
     
     public ArrayList<SanPhamChiTiet> HoaDonCho(String MaHD){
